Use floats for vertices corrdinates

+++ Project-Final/Headers/DataTypes.h
@@ -7,8 +7,8 @@
 
 struct VERTEX_T
 {
-    int32_t     X;
-    int32_t     Y;
+    float       X;
+    float       Y;
 };
 
 typedef std::vector<VERTEX_T>   VERTICES_T;
